Reset does not clear local/remote settings
