+++ linkup-cli/src/reset.rs
@@ -1,6 +1,11 @@
-use crate::{start, stop, CliError};
+use crate::{
+    background_booting::boot_background_services, start::get_state, stop::shutdown, CliError,
+};
 
-pub fn reset(config_arg: Option<String>) -> Result<(), CliError> {
-    stop()?;
-    start(config_arg)
+pub fn reset() -> Result<(), CliError> {
+    // Ensure there is some kind of state from before, otherwise reset doesn't make sense
+    get_state()?;
+
+    shutdown()?;
+    boot_background_services()
 }
